Fix natural.size in python 3

+++ natural/number.py
@@ -63,7 +63,7 @@ def ordinal(value):
     '''
 
     try:
-        value = long(value)
+        value = int(value)
     except (TypeError, ValueError):
         raise ValueError
 
@@ -152,7 +152,7 @@ def word(value, digits=2):
     decimal_point = convention['decimal_point']
     decimal_zero = re.compile(r'%s0+' % re.escape(decimal_point))
     prefix = value < 0 and u'-' or u''
-    value = abs(long(value))
+    value = abs(int(value))
     if value < 1000:
         return u''.join([
             prefix,
